feat: separate meta agent props and config interface
