+++ src/MetaAgent.ts
@@ -1,113 +1,7 @@
 import { IAgentPrompt } from "./structures/IAgentPrompt";
 import { IAgentTokenUsage } from "./structures/IAgentTokenUsage";
 import { IAgentEvent } from "./structures/IAgentEvent";
-import { IAgentController } from "./structures/IAgentController";
-import { IAgentProvider } from "./structures/IAgentProvider";
-import { Primitive } from "typia";
-
-export namespace MetaAgent {
-  export type Source =
-    | "initialize"
-    | "select"
-    | "cancel"
-    | "execute"
-    | "describe";
-  export type IProps = {
-    controllers: IAgentController[];
-    provider: IAgentProvider;
-    config?: IConfig;
-    histories?: Primitive<IAgentPrompt>[];
-  };
-
-  export interface IConfig {
-    /**
-     * Locale of the A.I. chatbot.
-     *
-     * If you configure this property, the A.I. chatbot will conversate with
-     * the given locale. You can get the locale value by
-     *
-     * - Browser: `navigator.language`
-     * - NodeJS: `process.env.LANG.split(".")[0]`
-     *
-     * @default your_locale
-     */
-    locale?: string;
-
-    /**
-     * Timezone of the A.I. chatbot.
-     *
-     * If you configure this property, the A.I. chatbot will consider the
-     * given timezone. You can get the timezone value by
-     * `Intl.DateTimeFormat().resolvedOptions().timeZone`.
-     *
-     * @default your_timezone
-     */
-    timezone?: string;
-
-    /**
-     * Retry count.
-     *
-     * If LLM function calling composed arguments are invalid,
-     * the A.I. chatbot will retry to call the function with
-     * the modified arguments.
-     *
-     * By the way, if you configure it to 0 or 1, the A.I. chatbot
-     * will not retry the LLM function calling for correcting the
-     * arguments.
-     *
-     * @default 3
-     */
-    retry?: number;
-
-    /**
-     * Capacity of the LLM function selecting.
-     *
-     * When the A.I. chatbot selects a proper function to call, if the
-     * number of functions registered in the
-     * {@link INestiaAgentProps.applications} is too much greater,
-     * the A.I. chatbot often fallen into the hallucination.
-     *
-     * In that case, if you configure this property value, `NestiaChatAgent`
-     * will divide the functions into the several groups with the configured
-     * capacity and select proper functions to call by operating the multiple
-     * LLM function selecting agents parallelly.
-     *
-     * @default 0
-     */
-    capacity?: number;
-
-    /**
-     * Eliticism for the LLM function selecting.
-     *
-     * If you configure {@link capacity}, the A.I. chatbot will complete
-     * the candidate functions to call which are selected by the multiple
-     * LLM function selecting agents.
-     *
-     * Otherwise you configure this property as `false`, the A.I. chatbot
-     * will not complete the candidate functions to call and just accept
-     * every candidate functions to call which are selected by the multiple
-     * LLM function selecting agents.
-     *
-     * @default true
-     */
-    eliticism?: boolean;
-
-    /**
-     * System prompt messages.
-     *
-     * System prompt messages if you want to customize the system prompt
-     * messages for each situation.
-     */
-    systemPrompt?: {
-      common?: (config?: IConfig) => string;
-      initialize?: (histories: IAgentPrompt[]) => string;
-      select?: (histories: IAgentPrompt[]) => string;
-      cancel?: (histories: IAgentPrompt[]) => string;
-      execute?: (histories: IAgentPrompt[]) => string;
-      describe?: (histories: IAgentPrompt.IExecute[]) => string;
-    };
-  }
-}
+import { IAgentProps } from "./structures/IAgentProps";
 
 export class MetaAgent {
   /* -----------------------------------------------------------
@@ -118,7 +12,7 @@ export class MetaAgent {
    *
    * @param props Properties to construct the agent
    */
-  public constructor(private readonly props: MetaAgent.IProps) {
+  public constructor(private readonly props: IAgentProps) {
     throw Error("Not implemented");
   }
